Search Contributions Cleanup (#165538)
Major cleanup on search contribs/actions, namely:
- converting all `Action` and related registrations to `Action2`
- Converting the `Actionbar` that appears on hover of a search result to `MenuWorkbenchToolBar`
- Overall organization
- Converting any `ICommandHandler` to functions if they don't need to be `ICommandHandler`
- Adding titles to all search commands
