+++ src/vs/workbench/contrib/search/browser/searchActionsSymbol.ts
@@ -1,0 +1,48 @@
+/*---------------------------------------------------------------------------------------------
+ *  Copyright (c) Microsoft Corporation. All rights reserved.
+ *  Licensed under the MIT License. See License.txt in the project root for license information.
+ *--------------------------------------------------------------------------------------------*/
+
+import * as nls from 'vs/nls';
+import { ServicesAccessor } from 'vs/platform/instantiation/common/instantiation';
+import * as Constants from 'vs/workbench/contrib/search/common/constants';
+import { Action2, MenuId, registerAction2 } from 'vs/platform/actions/common/actions';
+import { KeybindingWeight } from 'vs/platform/keybinding/common/keybindingsRegistry';
+import { KeyCode, KeyMod } from 'vs/base/common/keyCodes';
+import { IQuickInputService } from 'vs/platform/quickinput/common/quickInput';
+
+//#region Actions
+registerAction2(class ShowAllSymbolsAction extends Action2 {
+
+	static readonly ID = 'workbench.action.showAllSymbols';
+	static readonly LABEL = nls.localize('showTriggerActions', "Go to Symbol in Workspace...");
+	static readonly ALL_SYMBOLS_PREFIX = '#';
+
+	constructor(
+	) {
+		super({
+			id: Constants.ShowAllSymbolsActionId,
+			title: {
+				value: nls.localize('showTriggerActions', "Go to Symbol in Workspace..."),
+				original: 'Go to Symbol in Workspace...',
+				mnemonicTitle: nls.localize({ key: 'miGotoSymbolInWorkspace', comment: ['&& denotes a mnemonic'] }, "Go to Symbol in &&Workspace...")
+			},
+			f1: true,
+			keybinding: {
+				weight: KeybindingWeight.WorkbenchContrib,
+				primary: KeyMod.CtrlCmd | KeyCode.KeyT
+			},
+			menu: {
+				id: MenuId.MenubarGoMenu,
+				group: '3_global_nav',
+				order: 2
+			}
+		});
+	}
+
+	override async run(accessor: ServicesAccessor): Promise<void> {
+		accessor.get(IQuickInputService).quickAccess.show(ShowAllSymbolsAction.ALL_SYMBOLS_PREFIX);
+	}
+});
+
+//#endregion
